requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
DateTime==5.5
distro==1.9.0
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.16.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-redis==0.2.2
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
ml_dtypes==0.5.1
numpy==2.3.0
openai==1.82.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
ply==3.11
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
redisvl==0.7.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
xxhash==3.5.0
zope.interface==7.2
zstandard==0.23.0
